requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
country-list==1.0.0
defusedxml==0.7.1
Django==4.1.2
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
multidict==6.0.2
pycodestyle==2.9.1
Pygments==2.13.0
PySocks==1.7.1
pytz==2022.4
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.6.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
urllib3==1.26.12
